Phil's attempt to clean up and re-org the code
I hope this was helpful. I got rid of the errors in vscode, but the app is still throwning errors when I run it. I'm thinking of creating a new branch off the master to start fresh :) Ping me on slack with your thougths.
I won't be able to look at this much more today or tomorrow though.